+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The main focus is on user-observable behavior of the engine.
 * Added support for sharing Java objects using the experimental Node.js Worker Threads API.
 * Added support for ScriptEngine `GLOBAL_SCOPE` bindings.
 * Made `Bindings` created by `ScriptEngine#createBindings()` implement `AutoCloseable` to allow closing the underlying `Context`.
+* Do not provide `Java` builtin object when Java interop is disabled.
 
 ## Version 1.0.0 RC10
 * Added support for `Array.prototype.{flat,flatMap}`, [a Stage 3 proposal](https://github.com/tc39/proposal-flatMap).