+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The main focus is on user-observable behavior of the engine.
 * Do not provide `Java` builtin object when Java interop is disabled.
 * Added option `js.print` to disable the `print` and `printErr` builtins (default: `true`).
 * Added option `js.load` to disable the `load` and `loadWithNewGlobal` builtins (default: `true`).
+* Added option `js.polyglot-builtin` to disable the `Polyglot` builtins (default: `true`).
 
 ## Version 1.0.0 RC10
 * Added support for `Array.prototype.{flat,flatMap}`, [a Stage 3 proposal](https://github.com/tc39/proposal-flatMap).