+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The main focus is on user-observable behavior of the engine.
 * Added option `js.experimental-array-prototype` that sets prototype of
   array-like non-JS objects (like `ProxyArray` or Java `List`) to `Array.prototype`.
   It is possible to use functions like `map` or `forEach` on these objects directly then.
+* Updated Node.js to version 10.15.0.
 
 ## Version 1.0.0 RC11
 * Graal.js only supports ECMAScript 5 (ES5) and newer, and enforces that rule.